Atualização

+++ del.php
@@ -1,12 +1,20 @@
 <?php
+session_start();
 require_once "vendor/autoload.php";
 
-$pessoa    = new App\Model\Pessoa();
-$pessoaDao = new App\Model\PessoaDaoMysql();
+$logar = new App\Model\Logar();
+$logarDao = new App\Model\LogarDaoMysql();
 
-if($_SERVER['REQUEST_METHOD'] == "GET") {
-    $pessoaDao->delete($_REQUEST['id']);
+if($logarDao->checkLogin()) {
 
-    die(header("Location: http://".$_SERVER['HTTP_HOST']."/crud"));
+    $pessoa    = new App\Model\Pessoa();
+    $pessoaDao = new App\Model\PessoaDaoMysql();
+
+    if($_SERVER['REQUEST_METHOD'] == "GET") {
+        $pessoaDao->delete($_REQUEST['id']);
+
+        die(header("Location: http://".$_SERVER['HTTP_HOST']."/crud"));
+    }
+} else {
+    die(header("Location: http://".$_SERVER['HTTP_HOST']."/crud/login"));
 }
-
